+++ lib/food.js
@@ -4,6 +4,7 @@ const $ = require('jquery')
 const FoodRemove = require('./food-remove')
 const FoodAdd = require('./food-add')
 const FoodTable = require('./food-table')
+require('./edit-food')
 
 const Food = class Food {
 
@@ -14,85 +15,10 @@ const Food = class Food {
       FoodRemove.remove()
     })
 
-    $('.foods').on('click', '.food-cell', (event) => {
-      const cellValue = $(event.target).text();
-      const td = event.target
-      $(td).html($('<input />',{'value' : cellValue}).val(cellValue).attr('id', cellValue))
-
-      const cell = $(td).children()[0]
-
-      cell.focus()
-
-      $('table').on('focusout', cell, (event) => {
-        this.editName()
-      })
-
-      $(cell).keypress( (event) => {
-        if(event.which == 13) {
-          this.editName()
-        }
-      });
-    })
-
-    $('.foods').on('click', '.calorie-cell', (event) => {
-      const cellValue = $(event.target).text();
-      $(event.target).html($('<input />',{'value' : cellValue}).val(cellValue).attr('id', cellValue))
-
-      $(`#${cellValue}`).focus()
-
-      $('table').on('focusout', `#${cellValue}`, (event) => {
-        this.editCalories()
-      })
-
-      $(`#${cellValue}`).keypress( (event) => {
-        if(event.which == 13) {
-          this.editCalories()
-        }
-      });
-    })
-
     $('#new-food').submit( (event) => FoodAdd.addFood())
 
     $('form').on('submit', (event) => event.preventDefault());
   }
-
-  editName() {
-    const toEdit = {name: $(event.target).val(), calories: $(event.target).parent().next().text()}
-
-    const tr = $(event.target).parent().parent()
-    const id = tr.data('food-id')
-
-    $.ajax({
-      url: API + `foods/${id}`,
-      method: 'PATCH',
-      data: toEdit
-    })
-    .then( (data) => {
-      this.updateCellName(data, tr)
-    })
-    .fail( (error) => {
-      console.error(error)
-    });
-  };
-
-  editCalories() {
-    const toEdit = {name: $(event.target).parent().prev().text(), calories: $(event.target).val()}
-
-    const tr = $(event.target).parent().parent()
-    const id = tr.data('food-id')
-
-    $.ajax({
-      url: API + `foods/${id}`,
-      method: 'PATCH',
-      data: toEdit
-    })
-    .then( (data) => {
-      this.updateCellCalories(data, tr)
-    })
-    .fail( (error) => {
-      console.error(error)
-    });
-  };
 
   searchFoods(){
     let searchName = $('#food-filter').val()
@@ -102,23 +28,11 @@ const Food = class Food {
       method: 'GET',
     })
     .then( (data) => {
-      this.makeFoodTable(data)
+      FoodTable.makeFoodTable(data)
     })
     .fail( (error) => {
       console.error(error)
     })
-  }
-
-  updateCellName (data, tr) {
-    const td = ($(tr).children()[0])
-    $(td).children()[0].remove()
-    $(td).text(data[0].name)
-  }
-
-  updateCellCalories( data, tr) {
-    const td = ($(tr).children()[1])
-    $(td).children()[0].remove()
-    $(td).text(data[0].calories)
   }
 };
